Utilised quota unit test valar

+++ database/migrations/2023_05_24_044101_create_directrecruitment_onboarding_agent_table.php
@@ -48,6 +48,8 @@ return new class extends Migration
      */
     public function down(): void
     {
+      if (DB::getDriverName() !== 'sqlite') {
         Schema::dropIfExists('directrecruitment_onboarding_agent');
+      }
     }
 };
