+++ database/migrations/2023_07_20_122901_create_worker_event_attachments_table.php
@@ -38,6 +38,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('worker_event_attachments');
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::dropIfExists('worker_event_attachments');
+        }
     }
 };
