+++ database/migrations/2023_08_02_080051_create_worker_bulk_upload_table.php
@@ -48,6 +48,8 @@ return new class extends Migration
      */
     public function down(): void
     {
+      if (DB::getDriverName() !== 'sqlite') {
         Schema::dropIfExists('worker_bulk_upload');
+      }
     }
 };
