+++ database/migrations/2023_08_14_021433_create_total_management_cost_management_attachments_table.php
@@ -40,6 +40,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('total_management_cost_management_attachments');
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::dropIfExists('total_management_cost_management_attachments');
+        }
     }
 };
