+++ database/migrations/2023_08_28_091347_create_e-contract_payroll_bulk_upload_table.php
@@ -38,6 +38,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('e-contract_payroll_bulk_upload');
+        if (DB::getDriverName() !== 'sqlite') {
+            Schema::dropIfExists('e-contract_payroll_bulk_upload');
+        }
     }
 };
